refactor submit function
