+++ basic-testing-frontend/src/util/numbers.js
@@ -1,4 +1,17 @@
+import { validateStringNotEmpty, validateNumber } from './validation.js';
+
 export function transformToNumber(value) {
   return +value;
+}
+
+export function cleanNumbers(numberValues) {
+  const numbers = [];
+  for (const numberInput of numberValues) {
+    validateStringNotEmpty(numberInput);
+    const number = transformToNumber(numberInput);
+    validateNumber(number);
+    numbers.push(number);
+  }
+  return numbers;
 }
